Add ght_remove_first
Adds new interface function to remove the first or the oldest entry
in the hash table. This can be used to make the hash table behave like
a queue and a hash table at the same time. For eg. it can be used to
speed-up lookup is we are sure that the entry we're looking for is the
first entry in the table. If it is not, then we can proceed to
perform the hash based lookup.

Since the remove_from_chain function requires the index of the bucket
in which the entry resides, this patch also adds a new member to the
entry structure to store such an index. This index is updated at the
time of insertion or re-hashing.