Add support for RPC callback_info entry cache.
Use ght_remove_first function of the hash table to peek
at the first or the oldest entry in the hash table to see if can
get the RPC request's callback structure without performing a hash
table lookup. If the oldest entry is not the one we're looking for,
then performs the hash table lookup.